+++ app/src/main/java/com/app/music_app/logic/exercises/logic/AbstractExercise.kt
@@ -1,4 +1,4 @@
-package com.app.music_app.exercises.logic
+package com.app.music_app.logic.exercises.logic
 
 import androidx.compose.runtime.Composable
