+++ app/src/main/java/com/app/music_app/view/screens/TaskDifficultyScreen.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.app.music_app.exercises.difficulty.DifficultyInfo
+import com.app.music_app.logic.exercises.difficulty.DifficultyInfo
 import com.app.music_app.view.app_theme.AppTheme
 import com.example.android_app.R
